--- a/doc/etc/pgagroal.conf
+++ b/doc/etc/pgagroal.conf
@@ -10,6 +10,7 @@ max_connections = 100
 idle_timeout = 600
 validation = off
 unix_socket_dir = /tmp/
+ev_backend = auto
 
 [primary]
 host = localhost