… README.md
… cmake/FindLibev.cmake
… doc/ARCHITECTURE.md
… doc/CONFIGURATION.md
… doc/DEVELOPERS.md
… doc/man/pgagroal.conf.5.rst
… doc/manual/02-installation.md
… doc/manual/99-references.md
… doc/manual/dev-02-architecture.md
… doc/manual/dev-06-eventloop.md
… doc/manual/user-02-configuration.md
… src/CMakeLists.txt
… src/libpgagroal/configuration.c
… src/libpgagroal/ev.c
… src/libpgagroal/logging.c
… src/libpgagroal/pipeline_perf.c
… src/libpgagroal/pipeline_session.c
… src/libpgagroal/pipeline_transaction.c
… src/libpgagroal/utils.c
… src/libpgagroal/worker.c
… src/main.c
… src/vault.c