Introduce ev.h and ev.c, establishing the
foundation for the new custom event loop,
`pgagroal_ev`.

Replace previous dependencies on libev with the
custom event loop.

Implement support for io_uring with a fallback to
epoll if io_uring is unavailable.

Known Bugs
==========
- https://github.com/agroal/pgagroal/pull/456#issuecomment-2747668910
(not properly tested, might even exist upstream)

TODO
=====
- [x] Configuration reload issue;
- [x] Cleanup and write documentation;
- [x] CI: improve CI, use the pgmoneta example (have test scripts for signal handling and io);
- [x] kqueue: fix;
- [x] https://github.com/agroal/pgagroal/issues/503.

- [x] modify signal handling for io_uring (use libc handlers);
- [x] io_uring: use io_uring to read messages outside the ev loop: this is not an advantage as the other messages are exchanged in serial way
- [x] io_uring: find a way to get pdus read async (not viable)
- [x] io_uring: research how to tune io_uring for pgagroal use case: maybe test with multishot receive;


Changelog
=========

https://github.com/agroal/pgagroal/commit/d981313c2648eabd4bc524f71949f7e2213f176c
(2025-05-15):
- Add guard check in `pgagroal_event_loop_break`
- Rename ev_type -> event_type
- Create developer manual

https://github.com/agroal/pgagroal/commit/35b0a66380fcea11f78e5ae221a8f696df670bcc
(2025-05-11):
If the worker fails to create the loop, the loop will result in a null ptr dereference when destroyed. Fix that by moving up the destroyer of the loop to right after the loop exits.

https://github.com/agroal/pgagroal/commit/439fa05ca2a666e9c0434e5a40f62a951e2c8bf0
(2025-04-28):
Just updates based on the comments. I also did some tests with experimental features. These could be turned on by setting their flags to 1 on ev.h. They are still not correct and considering my tests I'm not even sure they offer benefits yet.

[290a4f0](https://github.com/agroal/pgagroal/pull/456/commits/290a4f0ecdb0443dc3acd4c2b2d26ef9edf8b221)
(2025-03-25):
ev_backend does not accept reload change, fix this.

[271ed06](https://github.com/agroal/pgagroal/pull/456/commits/271ed0687f5b67da72425843180b5905d48d5c54) (2025-03-25):
New CI to integrate test suite

[d73cafa](https://github.com/agroal/pgagroal/commit/d73cafac7a7eb5aee8f8d4a1e40e3ebb3294c5de)(2025-03-23):
Fix kqueue, creation of test suite, modify documentation of ARCHITECTURE.md

b18700ea96a68c83d5a4b079250fb73691e953dc (2025-03-17):
- [#503] pool.c: Fix bug that makes the database not release connections
    Fix this by sending 'X' to the database server inside
    pgagroal_kill_connection.

1076ef9061df9247014cb08562e24542f16fd7f3 (2025-03-17):
- ev.c: Implement io_uring infra for other IO operations and improve
    message handling
    io_uring functions for other IO operations (send/recv) was implemented,
    but is not being used because the advantage is questionable because
    other communication happens serially. Therefore it is not integrated
    yet. While we are at it, improve other parts of event handling for send
    and recv of io_uring messages. Also fix minor bug present upstream.

0a7de8dea7fb (2025-03-03):
- ev.c: added io_uring operations for recv-send in performance pipeline  
  Implemented an io_uring–specific recv-send compound operation to improve TPS.  
  Also modified variable and function names, and removed deprecated, inefficient, and unused code.

06028955a48e (2025-02-28):
- ev.c: Do some tuning and general cleanup

2fe657120dc4 (2025-02-28):
- configuration.c: Fix configuration of EV_BACKEND when invalid  
  Now sets config->ev_backend to EV_BACKEND_AUTO if the configuration is invalid.

f3318314a7fc (2025-02-27):
- ev.c: Improve signal handling for io_uring  
  Replaced the custom signal handler implementation with the glibc default.

09ac4bebb602 (2025-02-27):
- ev.c: Add some tuning options for io_uring and code cleanup

bcc24686516f (2025-02-27):
- libpgagroal/ev.c: improve debug/error messages

63b8ec5adb48 (2025-02-24):
- Fix "hang" after running pgbench, caused by using both
read_message directly on sockets and io_uring. For this, I
modified `worker.c`. I couldn't simply use pgagroal_ev_loop_destroy before the last exchange of messages between the worker and the server because the worker rely on allocd memory of which pgagroal_ev_loop_destroy gets rid. So epoll and io_uring will have to be treated differently until I find a sane way to do that.

2025-02-13:
- Fix ev_backend log in `configuration.c`. We cannot use
pgagroal_log_* functions during read configuration because the
log file is not set yet. Before this change, setting log type to
file leads to segmentation fault.

2025-01-18:
- delete unneeded file `bidi.c`.
- cleanup

2025-01-17:
- Fix regression added to the Debug build. `waitpid()`
inside `#if DEBUG` would block waiting for the running
child processes. Add WNOHANG to the function options so the
main process does not block.
- Remove '\n' from `pagroal_log_*()`
- Run uncrustify.sh
- Remove `printf()` debugging
- Modify CD/CI pipeline to improve function verify_running and
enable the CD/CI for kqueue, which was disabled.
- Change `pgagroal_log_error()` that informs errors in `kill()`
to `pgagroal_log_debug()`. I suppose these errors should happen
if pgagroal is not closely following up the death of its children.
Therefore, there is no need to flood the log with error messages.
Keep as debug because it is informative of something that could
be improved.

2025-01-16:
- Remove `pgagroal_ev_*_stop()` functions from main.c and vault.c
as these functions are not intended to be called from a child
process inheriting the ev loop. This does not make sense in
io_uring and for epoll and kqueue, just closing the fds is enough
as a precaution to prevent their use.
- Enhance debugging steps in `sigchld_handler()` to help debug
child processes termination.
- Add error handling to `kill()` to help debug child processes
termination.
- Allow return error on `pgagroal_ev_io_stop()` instead of exit
so the caller function can handle errors appropriately.
- Add a call to `setpgid()` when worker is created to ensure
child processes do not exit immediately if a SIGINT is issued
in the controlling terminal.
- Add a call to `pgagroal_ev_fork()` when a worker is created.

2025-01-13:
- Fix ASan report in `pgagroal_ev_io_stop()`