@jesperpedersen 2024-09-15
```
Run sudo mkdir -p /etc/pgagroal
kernel.io_uring_disabled = 0
pgagroal: PID file </tmp//pgagroal.2345.pid> exists, is there another instance running ?
```

Issue with shutdown ?