@decarv 2024-09-15
> ```
> pgagroal: PID file </tmp//pgagroal.2345.pid> exists, is there another instance running ?
> ```

Yes. I will investigate shutdown issues. However, from the CI output it could be anything, really.