@jesperpedersen 2024-09-15
Try and a TRACE log file for gcc and clang, and see if there are pointers there...