@decarv 2024-09-22
Please take a look at the latest commits. I decided to keep kqueue in this same PR for simplicity.