@jesperpedersen 2024-09-23
Just a quick pass on the first part of the patch in order to give some comments that are probably general for the entire patch