@fluca1978 2024-09-23
I still have issues with ipv6 listening addresses: pgagroal is unresponsive if `listen_address = *`.
Changing `listen_address` to `localhost` allows me to test against pgbench, and so far it seems a little faster than the master branch, I'm still testing.
`pgagroal-cli shutdown` works fine, while stopping with `SIGINT` causes the pid file to stale.