@decarv 2024-09-24
> I still have issues with ipv6 listening addresses: pgagroal is unresponsive if `listen_address = *`. Changing `listen_address` to `localhost` allows me to test against pgbench, and so far it seems a little faster than the master branch, I'm still testing.

I will investigate that further to understand what is happening. I don't see how ev.c is doing this (although it has to be it).

> `pgagroal-cli shutdown` works fine, while stopping with `SIGINT` causes the pid file to stale.
Thank you. Fixed.