@decarv 2024-09-30
> > I still have issues with ipv6 listening addresses: pgagroal is unresponsive if `listen_address = *`. Changing `listen_address` to `localhost` allows me to test against pgbench, and so far it seems a little faster than the master branch, I'm still testing.
> 
> I will investigate that further to understand what is happening. I don't see how ev.c is doing this (although it has to be it).
> 
> > `pgagroal-cli shutdown` works fine, while stopping with `SIGINT` causes the pid file to stale.
> 
> Thank you. Fixed. Already on the next commit.

@fluca1978 can you please confirm you do not see this behavior in master branch? I also get
> Thank you. Fixed. Already on the next commit.

@fluca1978 I think i have identified the issue. Can you please confirm you do not see this behavior in master branch? I also get an error in master branch... If you confirm that you can see a problem in master branch I will open another issue and fix it. Otherwise I will have to keep looking.