@fluca1978 2024-11-13
Testing commit e014d881e7983b24ef88d22cb2e1e0d168a49f03 on FreeBSD 14.1.
Database limits are configured as follows:

```
% cat /usr/local/etc/pgagroal/pgagroal_databases.conf
# DATABASE USER MAX INITIAL MIN
 pgbench pgbench 10 5  5
```

With `ev_backend = io_uring` (note this should not work, according to  me):

```
% pgbench -T 60  -h localhost -p 54322 -P 10 -c 8 -n -U pgbench pgbench
pgbench (17.0)
progress: 10.0 s, 742.1 tps, lat 8.594 ms stddev 17.618, 0 failed
progress: 20.0 s, 693.2 tps, lat 9.373 ms stddev 9.618, 0 failed
progress: 30.0 s, 687.4 tps, lat 9.443 ms stddev 9.341, 0 failed
progress: 40.0 s, 686.4 tps, lat 9.523 ms stddev 9.598, 0 failed
progress: 50.0 s, 683.4 tps, lat 9.495 ms stddev 9.566, 0 failed
progress: 60.0 s, 683.0 tps, lat 9.603 ms stddev 9.589, 0 failed
transaction type: <builtin: TPC-B (sort of)>
scaling factor: 1000
query mode: simple
number of clients: 8
number of threads: 1
maximum number of tries: 1
duration: 60 s
number of transactions actually processed: 41778
number of failed transactions: 0 (0.000%)
latency average = 9.333 ms
latency stddev = 11.416 ms
initial connection time = 339.957 ms
tps = 699.899185 (without initial connection time)
```

everything seems to work fine, and I'm able to launch the same test over and over and it runs.
However, this is using `io_uring` as `ev_backend`, which is "strange".

Configuring it with `kqueue`


```
% pgagroal
-> WARN  configuration.c:1774 pgagroal: Remote management enabled, but no admins are defined
-> DEBUG ev.c:562 Available ev backends: kqueue
-> DEBUG ev.c:577 Selected backend: 'kqueue'
-> INFO  main.c:1138 pgagroal: 2.0.0 started on localhost:54322
...

and the initial situation appears to be good (5 initial connections):

```
pgbench=# select pid, usename, datname, query from pg_stat_activity where datname= current_database();
 pid  | usename  | datname |                                            query                                             
-------+----------+---------+----------------------------------------------------------------------------------------------
 15072 | postgres | pgbench | select pid, usename, datname, query from pg_stat_activity where datname= current_database();
 15182 | pgbench  | pgbench | DISCARD ALL;
 15183 | pgbench  | pgbench | DISCARD ALL;
 15184 | pgbench  | pgbench | DISCARD ALL;
 15185 | pgbench  | pgbench | DISCARD ALL;
 15186 | pgbench  | pgbench | DISCARD ALL;
(6 rows)

```

but it is not possible to run the test:

```
% pgbench -T 60  -h localhost -p 54322 -P 10 -c 8 -n -U pgbench pgbench
pgbench (17.0)
pgbench: error: could not count number of branches: server closed the connection unexpectedly
        This probably means the server terminated abnormally
        before or while processing the request.

% pgbench -T 60  -h localhost -p 54322 -P 10 -c 2 -n -U pgbench pgbench
pgbench (17.0)
pgbench: error: could not count number of branches: server closed the connection unexpectedly
        This probably means the server terminated abnormally
        before or while processing the request.


% pgbench -T 60  -h localhost -p 54322 -P 10  -n -U pgbench pgbench 
pgbench (17.0)
pgbench: error: could not count number of branches: server closed the connection unexpectedly
        This probably means the server terminated abnormally
        before or while processing the request.

```

So I tried a single connection, but the connection is immediatly discarded:

```
% psql -h localhost -U pgbench -p 54322 -c 'select current_database();'  pgbench
server closed the connection unexpectedly
        This probably means the server terminated abnormally
        before or while processing the request.
connection to server was lost
```

and this is the log from pgagroal:

```
-> ERROR ev.c:1648 __kqueue_io_stop: kevent delete failed
-> ERROR ev.c:1648 __kqueue_io_stop: kevent delete failed
-> ERROR ev.c:1648 __kqueue_io_stop: kevent delete failed
-> ERROR ev.c:1648 __kqueue_io_stop: kevent delete failed
-> DEBUG security.c:237 SSL request from client: 13
-> DEBUG security.c:389 authenticate: getting pooled connection
-> DEBUG security.c:1666 client_trust 13 1
-> DEBUG security.c:402 authenticate: got pooled connection (1)
-> DEBUG security.c:428 authenticate: SUCCESS
-> DEBUG worker.c:96 pgagroal_worker: Slot 1 (13 -> 15)
-> INFO  worker.c:103 connect: user=pgbench database=pgbench address=d00::503a:7127:800:0
-> DEBUG pool.c:1227 pgagroal_pool_status: 1/20
-> DEBUG pool.c:1358 pgagroal_pool_status: State: NOTINIT
-> DEBUG pool.c:1359                       Slot: 0
-> DEBUG pool.c:1360                       FD: -1
-> DEBUG pool.c:1389 pgagroal_pool_status: State: IN_USE
-> DEBUG pool.c:1390                       Slot: 1
-> DEBUG pool.c:1391                       Server: 0
-> DEBUG pool.c:1392                       User: pgbench
-> DEBUG pool.c:1393                       Database: pgbench
-> DEBUG pool.c:1394                       AppName: psql
-> DEBUG pool.c:1395                       Rule: 0
-> DEBUG pool.c:1396                       Start: Wed Nov 13 11:38:05 2024
-> DEBUG pool.c:1397                       Time: Wed Nov 13 11:39:39 2024
-> DEBUG pool.c:1398                       FD: 15
-> DEBUG pool.c:1368 pgagroal_pool_status: State: FREE
-> DEBUG pool.c:1369                       Slot: 2
-> DEBUG pool.c:1370                       Server: 0
-> DEBUG pool.c:1371                       User: pgbench
-> DEBUG pool.c:1372                       Database: pgbench
-> DEBUG pool.c:1373                       AppName: 
-> DEBUG pool.c:1374                       Rule: 0
-> DEBUG pool.c:1375                       Start: Wed Nov 13 11:38:05 2024
-> DEBUG pool.c:1376                       Time: Wed Nov 13 11:38:05 2024
-> DEBUG pool.c:1377                       FD: 16
-> DEBUG pool.c:1368 pgagroal_pool_status: State: FREE
-> DEBUG pool.c:1369                       Slot: 3
-> DEBUG pool.c:1370                       Server: 0
-> DEBUG pool.c:1371                       User: pgbench
-> DEBUG pool.c:1372                       Database: pgbench
-> DEBUG pool.c:1373                       AppName: 
-> DEBUG pool.c:1374                       Rule: 0
-> DEBUG pool.c:1375                       Start: Wed Nov 13 11:38:05 2024
-> DEBUG pool.c:1376                       Time: Wed Nov 13 11:38:05 2024
-> DEBUG pool.c:1377                       FD: 17
-> DEBUG pool.c:1368 pgagroal_pool_status: State: FREE
-> DEBUG pool.c:1369                       Slot: 4
-> DEBUG pool.c:1370                       Server: 0
-> DEBUG pool.c:1371                       User: pgbench
-> DEBUG pool.c:1372                       Database: pgbench
-> DEBUG pool.c:1373                       AppName: 
-> DEBUG pool.c:1374                       Rule: 0
-> DEBUG pool.c:1375                       Start: Wed Nov 13 11:38:05 2024
-> DEBUG pool.c:1376                       Time: Wed Nov 13 11:38:05 2024
-> DEBUG pool.c:1377                       FD: 18
-> DEBUG pool.c:1358 pgagroal_pool_status: State: NOTINIT
-> DEBUG pool.c:1359                       Slot: 5
-> DEBUG pool.c:1360                       FD: -1
-> DEBUG pool.c:1358 pgagroal_pool_status: State: NOTINIT
-> DEBUG pool.c:1359                       Slot: 6
-> DEBUG pool.c:1360                       FD: -1
-> DEBUG pool.c:1358 pgagroal_pool_status: State: NOTINIT
-> DEBUG pool.c:1359                       Slot: 7
-> DEBUG pool.c:1360                       FD: -1
-> DEBUG pool.c:1358 pgagroal_pool_status: State: NOTINIT
-> DEBUG pool.c:1359                       Slot: 8
-> DEBUG pool.c:1360                       FD: -1
-> DEBUG pool.c:1358 pgagroal_pool_status: State: NOTINIT
-> DEBUG pool.c:1359                       Slot: 9
-> DEBUG pool.c:1360                       FD: -1
-> DEBUG pool.c:1358 pgagroal_pool_status: State: NOTINIT
-> DEBUG pool.c:1359                       Slot: 10
-> DEBUG pool.c:1360                       FD: -1
-> DEBUG pool.c:1358 pgagroal_pool_status: State: NOTINIT
-> DEBUG pool.c:1359                       Slot: 11
-> DEBUG pool.c:1360                       FD: -1
-> DEBUG pool.c:1358 pgagroal_pool_status: State: NOTINIT
-> DEBUG pool.c:1359                       Slot: 12
-> DEBUG pool.c:1360                       FD: -1
-> DEBUG pool.c:1358 pgagroal_pool_status: State: NOTINIT
-> DEBUG pool.c:1359                       Slot: 13
-> DEBUG pool.c:1360                       FD: -1
-> DEBUG pool.c:1358 pgagroal_pool_status: State: NOTINIT
-> DEBUG pool.c:1359                       Slot: 14
-> DEBUG pool.c:1360                       FD: -1
-> DEBUG pool.c:1358 pgagroal_pool_status: State: NOTINIT
-> DEBUG pool.c:1359                       Slot: 15
-> DEBUG pool.c:1360                       FD: -1
-> DEBUG pool.c:1358 pgagroal_pool_status: State: NOTINIT
-> DEBUG pool.c:1359                       Slot: 16
-> DEBUG pool.c:1360                       FD: -1
-> DEBUG pool.c:1358 pgagroal_pool_status: State: NOTINIT
-> DEBUG pool.c:1359                       Slot: 17
-> DEBUG pool.c:1360                       FD: -1
-> DEBUG pool.c:1358 pgagroal_pool_status: State: NOTINIT
-> DEBUG pool.c:1359                       Slot: 18
-> DEBUG pool.c:1360                       FD: -1
-> DEBUG pool.c:1358 pgagroal_pool_status: State: NOTINIT
-> DEBUG pool.c:1359                       Slot: 19
-> DEBUG pool.c:1360                       FD: -1
-> DEBUG ev.c:562 Available ev backends: kqueue
-> DEBUG ev.c:577 Selected backend: 'kqueue'
-> WARN  pipeline_perf.c:190 [C] Client error (slot 1 database pgbench user pgbench): Resource temporarily unavailable (socket 13 status 2)
-> INFO  message.c:1203 Message is NULL
-> INFO  worker.c:206 disconnect: user=pgbench database=pgbench address=d00::503a:7127:800:0
-> DEBUG pool.c:428 pgagroal_return_connection: Slot 1 FD 15
-> DEBUG worker.c:291 client disconnect: 13
-> DEBUG pool.c:1227 pgagroal_pool_status: 0/20
-> DEBUG pool.c:1358 pgagroal_pool_status: State: NOTINIT
-> DEBUG pool.c:1359                       Slot: 0
-> DEBUG pool.c:1360                       FD: -1
-> DEBUG pool.c:1368 pgagroal_pool_status: State: FREE
-> DEBUG pool.c:1369                       Slot: 1
-> DEBUG pool.c:1370                       Server: 0
-> DEBUG pool.c:1371                       User: pgbench
-> DEBUG pool.c:1372                       Database: pgbench
-> DEBUG pool.c:1373                       AppName: 
-> DEBUG pool.c:1374                       Rule: 0
-> DEBUG pool.c:1375                       Start: Wed Nov 13 11:38:05 2024
-> DEBUG pool.c:1376                       Time: Wed Nov 13 11:39:39 2024
-> DEBUG pool.c:1377                       FD: 15
-> DEBUG pool.c:1368 pgagroal_pool_status: State: FREE
-> DEBUG pool.c:1369                       Slot: 2
-> DEBUG pool.c:1370                       Server: 0
-> DEBUG pool.c:1371                       User: pgbench
-> DEBUG pool.c:1372                       Database: pgbench
-> DEBUG pool.c:1373                       AppName: 
-> DEBUG pool.c:1374                       Rule: 0
-> DEBUG pool.c:1375                       Start: Wed Nov 13 11:38:05 2024
-> DEBUG pool.c:1376                       Time: Wed Nov 13 11:38:05 2024
-> DEBUG pool.c:1377                       FD: 16
-> DEBUG pool.c:1368 pgagroal_pool_status: State: FREE
-> DEBUG pool.c:1369                       Slot: 3
-> DEBUG pool.c:1370                       Server: 0
-> DEBUG pool.c:1371                       User: pgbench
-> DEBUG pool.c:1372                       Database: pgbench
-> DEBUG pool.c:1373                       AppName: 
-> DEBUG pool.c:1374                       Rule: 0
-> DEBUG pool.c:1375                       Start: Wed Nov 13 11:38:05 2024
-> DEBUG pool.c:1376                       Time: Wed Nov 13 11:38:05 2024
-> DEBUG pool.c:1377                       FD: 17
-> DEBUG pool.c:1368 pgagroal_pool_status: State: FREE
-> DEBUG pool.c:1369                       Slot: 4
-> DEBUG pool.c:1370                       Server: 0
-> DEBUG pool.c:1371                       User: pgbench
-> DEBUG pool.c:1372                       Database: pgbench
-> DEBUG pool.c:1373                       AppName: 
-> DEBUG pool.c:1374                       Rule: 0
-> DEBUG pool.c:1375                       Start: Wed Nov 13 11:38:05 2024
-> DEBUG pool.c:1376                       Time: Wed Nov 13 11:38:05 2024
-> DEBUG pool.c:1377                       FD: 18
-> DEBUG pool.c:1358 pgagroal_pool_status: State: NOTINIT
-> DEBUG pool.c:1359                       Slot: 5
-> DEBUG pool.c:1360                       FD: -1
-> DEBUG pool.c:1358 pgagroal_pool_status: State: NOTINIT
-> DEBUG pool.c:1359                       Slot: 6
-> DEBUG pool.c:1360                       FD: -1
-> DEBUG pool.c:1358 pgagroal_pool_status: State: NOTINIT
-> DEBUG pool.c:1359                       Slot: 7
-> DEBUG pool.c:1360                       FD: -1
-> DEBUG pool.c:1358 pgagroal_pool_status: State: NOTINIT
-> DEBUG pool.c:1359                       Slot: 8
-> DEBUG pool.c:1360                       FD: -1
-> DEBUG pool.c:1358 pgagroal_pool_status: State: NOTINIT
-> DEBUG pool.c:1359                       Slot: 9
-> DEBUG pool.c:1360                       FD: -1
-> DEBUG pool.c:1358 pgagroal_pool_status: State: NOTINIT
-> DEBUG pool.c:1359                       Slot: 10
-> DEBUG pool.c:1360                       FD: -1
-> DEBUG pool.c:1358 pgagroal_pool_status: State: NOTINIT
-> DEBUG pool.c:1359                       Slot: 11
-> DEBUG pool.c:1360                       FD: -1
-> DEBUG pool.c:1358 pgagroal_pool_status: State: NOTINIT
-> DEBUG pool.c:1359                       Slot: 12
-> DEBUG pool.c:1360                       FD: -1
-> DEBUG pool.c:1358 pgagroal_pool_status: State: NOTINIT
-> DEBUG pool.c:1359                       Slot: 13
-> DEBUG pool.c:1360                       FD: -1
-> DEBUG pool.c:1358 pgagroal_pool_status: State: NOTINIT
-> DEBUG pool.c:1359                       Slot: 14
-> DEBUG pool.c:1360                       FD: -1
-> DEBUG pool.c:1358 pgagroal_pool_status: State: NOTINIT
-> DEBUG pool.c:1359                       Slot: 15
-> DEBUG pool.c:1360                       FD: -1
-> DEBUG pool.c:1358 pgagroal_pool_status: State: NOTINIT
-> DEBUG pool.c:1359                       Slot: 16
-> DEBUG pool.c:1360                       FD: -1
-> DEBUG pool.c:1358 pgagroal_pool_status: State: NOTINIT
-> DEBUG pool.c:1359                       Slot: 17
-> DEBUG pool.c:1360                       FD: -1
-> DEBUG pool.c:1358 pgagroal_pool_status: State: NOTINIT
-> DEBUG pool.c:1359                       Slot: 18
-> DEBUG pool.c:1360                       FD: -1
-> DEBUG pool.c:1358 pgagroal_pool_status: State: NOTINIT
-> DEBUG pool.c:1359                       Slot: 19
-> DEBUG pool.c:1360                       FD: -1
-> DEBUG worker.c:299 After client: PID 15359 Slot 1 (3)
-> DEBUG main.c:1972 pgagroal: Transfer return connection: Slot 1
-> DEBUG main.c:2019 pgagroal: Transfer client done
-> DEBUG main.c:2029 pgagroal: Transfer client done: PID 15359
```