@decarv 2025-03-05
Pushed new patches that modify functionality of performance pipeline. The changes can be easily propagated to the other pipelines, I will do it this week. After this, I will only have kqueue left.

I have also started adding tests. Please take a look and make suggestions (they are not ready yet).

At last, some preliminary results for pipeline performance:

```

TPS Table:
+---------+-------------------+---------------+------------------+
| Clients | master branch TPS | new epoll TPS | new io uring TPS |
+---------+-------------------+---------------+------------------+
|    5    |     3910.34       |    3973.12    |     3412.06      |
|   20    |     4403.88       |    5241.68    |     6917.67      |
|   50    |     4540.13       |    5410.73    |     6433.11      |
|   90    |     5099.07       |    6224.38    |     7520.74      |
+---------+-------------------+---------------+------------------+


Average Latency (ms):
+---------+----------------------------+------------------------+---------------------------+
| Clients | master branch Latency (ms) | new epoll Latency (ms) | new io uring Latency (ms) |
+---------+----------------------------+------------------------+---------------------------+
|    5    |         1.279              |         1.258          |         1.465             |
|   20    |         4.541              |         3.816          |         2.891             |
|   50    |         11.013             |         9.241          |         7.772             |
|   90    |         17.650             |         14.459         |         11.967            |
+---------+----------------------------+------------------------+---------------------------+


Initial Connection Time (ms):
+---------+----------------------------------+------------------------------+--------------------------------+
| Clients | master branch Initial Conn (ms)  | new epoll Initial Conn (ms)  | new io uring Initial Conn (ms) |
+---------+----------------------------------+------------------------------+--------------------------------+
|    5    |           1.713                |           2.720              |           1.672                |
|   20    |           3.867                |           2.700              |           3.018                |
|   50    |           6.487                |           5.934              |           7.383                |
|   90    |          15.787                |          24.144              |          25.931                |
+---------+----------------------------------+------------------------------+--------------------------------+

```